requirements.txt
# Core dependencies
langchain>=0.1.0
requests>=2.31.0
pyyaml>=6.0

# Data processing and ML
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.24.0
joblib>=1.3.0

# Utilities
python-dateutil>=2.8.2
colorama>=0.4.6

# LangChain libraries
langchain-community>=0.0.21
langchain-core>=0.1.26

# RAG and ML dependencies 
tqdm==4.66.2

# After installing the requirements, you'll also need to download the spaCy model:
# python -m spacy download en_core_web_sm 